Updates stream table to have additional data
In order to expand on the available data to get from streams, this
commit adds `twitch_vod_id`, `title`, and `muted_vod_segments` tracking
to the stream table. `muted_vod_segments` being its own table with a one
to many relation.

Additionally, to track this data a `nightly_checks` crate was made to go
over anything that needs to be compared against Twitch's servers such as
name changes or relevant vod data.

Lastly, the tracker's stream status update parsing was moved to the
websocket module for improved consistency on who handles what.