Implementes quality messages on report generator
In order to increase the quality of filtered messages in the filtered
chat ranking reports, this commit replaces emote filtered messages with
quality filtered messages. With this any messages that are a certain
percentage of emotes *or* contain less than 4 real words (not emotes)
are filtered in the filtered chat rankings.