Fix 404 errors and global emote list retrieval
In order to prevent the app from crashing at startup when a user doesn't
have a BTTV or FFZ account configured, this commit adds a check for the
method that fetches an emote list to ignore any 404 errors. In addition,
this commit moves the global emote retrieval to its own method instead
of attempting to fetch it for every user.